+++ mine/application.py
@@ -1,45 +1,53 @@
 """Data structures for application information."""
 
-from . import common
+import logging
+
 from .base import NameMixin
 
 import yorm
 
-log = common.logger(__name__)
+
+log = logging.getLogger(__name__)
 
 
 @yorm.attr(mac=yorm.converters.NoneString)
 @yorm.attr(windows=yorm.converters.NoneString)
 @yorm.attr(linux=yorm.converters.NoneString)
 class Versions(yorm.converters.AttributeDictionary):
+    """Dictionary of OS-specific application filenames."""
 
-    """A dictionary of OS-specific application filenames."""
 
-    def __init__(self, mac=None, windows=None, linux=None):
-        super().__init__()
-        self.mac = mac
-        self.windows = windows
-        self.linux = linux
+@yorm.attr(auto_queue=yorm.converters.Boolean)
+@yorm.attr(single_instance=yorm.converters.Boolean)
+class Properties(yorm.converters.AttributeDictionary):
+    """Dictionary of application management settings."""
 
 
 @yorm.attr(name=yorm.converters.String)
-@yorm.attr(queued=yorm.converters.Boolean)
+@yorm.attr(properties=Properties)
 @yorm.attr(versions=Versions)
 class Application(NameMixin, yorm.converters.AttributeDictionary):
+    """Dictionary of application information."""
 
-    """A dictionary of application information."""
+    # pylint: disable=E1101
 
-    def __init__(self, name, queued=False, filename=None):
+    def __init__(self, name, filename=None):
         super().__init__()
         self.name = name
-        self.queued = queued
         self.versions = Versions(mac=filename, windows=filename, linux=filename)
+
+    @property
+    def auto_queue(self):
+        return self.properties.auto_queue
+
+    @property
+    def no_wait(self):
+        return not self.properties.single_instance
 
 
 @yorm.attr(all=Application)
 class Applications(yorm.converters.SortedList):
-
-    """A list of monitored applications."""
+    """List of monitored applications."""
 
     def get(self, name):
         """Get the application with the given name."""
@@ -49,7 +57,7 @@ class Applications(yorm.converters.SortedList):
 
     def find(self, name):
         """Find the application with the given name, else None."""
-        log.debug("finding application for '%s'...", name)
+        log.debug("Finding application for '%s'...", name)
         for application in self:
             if application == name:
                 return application