+++ mine/computer.py
@@ -2,21 +2,20 @@
 
 import uuid
 import socket
-
+import logging
 
 import yorm
 
-from . import common
 from .base import NameMixin
 
-log = common.logger(__name__)
+
+log = logging.getLogger(__name__)
 
 
 @yorm.attr(name=yorm.converters.String)
 @yorm.attr(hostname=yorm.converters.String)
 @yorm.attr(address=yorm.converters.String)
 class Computer(NameMixin, yorm.converters.AttributeDictionary):
-
     """A dictionary of identifying computer information."""
 
     def __init__(self, name, hostname=None, address=None):
@@ -40,7 +39,6 @@ class Computer(NameMixin, yorm.converters.AttributeDictionary):
 
 @yorm.attr(all=Computer)
 class Computers(yorm.converters.SortedList):
-
     """A list of computers."""
 
     @property
@@ -56,14 +54,14 @@ class Computers(yorm.converters.SortedList):
 
     def find(self, name):
         """Find the computer with the given name, else None."""
-        log.debug("finding computer for '%s'...", name)
+        log.debug("Finding computer for '%s'...", name)
         for computer in self:
             if computer == name:
                 return computer
 
     def match(self, partial):
         """Find a computer with a similar name."""
-        log.debug("finding computer similar to '%s'...", partial)
+        log.debug("Finding computer similar to '%s'...", partial)
         for computer in self:
             if partial.lower() in computer.name.lower():
                 return computer
@@ -86,7 +84,7 @@ class Computers(yorm.converters.SortedList):
 
         # Else, this is a new computer
         this.name = self.generate_name(this)
-        log.debug("new computer: %s", this)
+        log.debug("New computer: %s", this)
         self.append(this)
         return this