+++ mine/config.py
@@ -1,19 +1,20 @@
 """Data structures for all settings."""
 
+import logging
+
 import yorm
 
-from . import common
 from .application import Applications
 from .computer import Computers
 
-log = common.logger(__name__)
+
+log = logging.getLogger(__name__)
 
 
 @yorm.attr(applications=Applications)
 @yorm.attr(computers=Computers)
 class ProgramConfig(yorm.converters.AttributeDictionary):
-
-    """A dictionary of program configuration settings."""
+    """Dictionary of program configuration settings."""
 
     def __init__(self):
         super().__init__()