+++ mine/manager.py
@@ -7,12 +7,12 @@ import glob
 import platform
 import functools
 import subprocess
+import logging
 
 import psutil
 
-from . import common
 
-log = common.logger(__name__)
+log = logging.getLogger(__name__)
 
 # TODO: delete this after implementing `BaseManager`
 # https://github.com/jacebrowning/mine/issues/8
@@ -26,14 +26,14 @@ def log_running(func):  # pragma: no cover (manual)
     @functools.wraps(func)
     def wrapped(self, application):
         """Wrapped method to log if an application is running."""
-        log.debug("determining if %s is running...", application)
+        log.debug("Determining if %s is running...", application)
         running = func(self, application)
         if running is None:
-            status = "untracked"
+            status = "Untracked"
         elif running:
-            status = "running"
+            status = "Running"
         else:
-            status = "not running"
+            status = "Not running"
         log.info("%s: %s", status, application)
         return running
     return wrapped
@@ -45,9 +45,9 @@ def log_starting(func):  # pragma: no cover (manual)
     @functools.wraps(func)
     def wrapped(self, application):
         """Wrapped method to log that an application is being started."""
-        log.info("starting %s...", application)
+        log.info("Starting %s...", application)
         result = func(self, application)
-        log.info("running: %s", application)
+        log.info("Running: %s", application)
         return result
     return wrapped
 
@@ -58,15 +58,14 @@ def log_stopping(func):  # pragma: no cover (manual)
     @functools.wraps(func)
     def wrapped(self, application):
         """Wrapped method to log that an application is being stopped."""
-        log.info("stopping %s...", application)
+        log.info("Stopping %s...", application)
         result = func(self, application)
-        log.info("not running: %s", application)
+        log.info("Not running: %s", application)
         return result
     return wrapped
 
 
 class BaseManager(metaclass=abc.ABCMeta):  # pragma: no cover (abstract)
-
     """Base application manager."""
 
     NAME = FRIENDLY = None
@@ -97,7 +96,7 @@ class BaseManager(metaclass=abc.ABCMeta):  # pragma: no cover (abstract)
     @staticmethod
     def _get_process(name):
         """Get a process whose executable path contains an app name."""
-        log.debug("searching for exe path containing '%s'...", name)
+        log.debug("Searching for exe path containing '%s'...", name)
 
         for process in psutil.process_iter():
             try:
@@ -108,11 +107,11 @@ class BaseManager(metaclass=abc.ABCMeta):  # pragma: no cover (abstract)
 
                 if name.lower() in parts:
                     if process.pid == os.getpid():
-                        log.debug("skipped current process: %s", command)
+                        log.debug("Skipped current process: %s", command)
                     elif process.status() == psutil.STATUS_ZOMBIE:
-                        log.debug("skipped zombie process: %s", command)
+                        log.debug("Skipped zombie process: %s", command)
                     else:
-                        log.debug("found matching process: %s", command)
+                        log.debug("Found matching process: %s", command)
                         return process
 
             except psutil.AccessDenied:
@@ -122,7 +121,6 @@ class BaseManager(metaclass=abc.ABCMeta):  # pragma: no cover (abstract)
 
 
 class LinuxManager(BaseManager):  # pragma: no cover (manual)
-
     """Application manager for Linux."""
 
     NAME = 'Linux'
@@ -145,12 +143,11 @@ class LinuxManager(BaseManager):  # pragma: no cover (manual)
             process.terminate()
 
     def launch(self, path):
-        log.info("opening %s...", path)
+        log.info("Opening %s...", path)
         return subprocess.call(['xdg-open', path]) == 0
 
 
 class MacManager(BaseManager):  # pragma: no cover (manual)
-
     """Application manager for OS X."""
 
     NAME = 'Darwin'
@@ -167,19 +164,21 @@ class MacManager(BaseManager):  # pragma: no cover (manual)
     @log_starting
     def start(self, application):
         name = application.versions.mac
-        if os.path.exists(name):
-            path = name
-        else:
-            path = os.path.join('/Applications', name)
-            if not os.path.exists(path):
-                pattern = os.path.join("/Applications/*", name)
-                log.debug("glob pattern: %s", pattern)
-                paths = glob.glob(pattern)
-                for path in paths:
-                    log.debug("match: %s", path)
-                assert paths, "not found: {}".format(application)
+        path = None
+        for base in (".",
+                     "~/Applications",
+                     "/Applications",
+                     "/Applications/*"):
+            pattern = os.path.expanduser(os.path.join(base, name))
+            log.debug("Glob pattern: %s", pattern)
+            paths = glob.glob(pattern)
+            if paths:
                 path = paths[0]
-        self._start_app(path)
+                log.debug("Match: %s", path)
+                break
+        else:
+            assert path, "Not found: {}".format(application)
+        return self._start_app(path)
 
     @log_stopping
     def stop(self, application):
@@ -203,7 +202,6 @@ class MacManager(BaseManager):  # pragma: no cover (manual)
 
 
 class WindowsManager(BaseManager):  # pragma: no cover (manual)
-
     """Application manager for Windows."""
 
     NAME = 'Windows'
@@ -226,7 +224,7 @@ class WindowsManager(BaseManager):  # pragma: no cover (manual)
 
 def get_manager(name=None):
     """Return an application manager for the current operating system."""
-    log.info("detecting the current system...")
+    log.info("Detecting the current system...")
     name = name or platform.system()
     if name == WindowsManager.NAME:
         manager = WindowsManager()
@@ -236,5 +234,5 @@ def get_manager(name=None):
         assert name == LinuxManager.NAME
         manager = LinuxManager()
 
-    log.info("current system: %s", manager)
+    log.info("Current system: %s", manager)
     return manager