+++ mine/timestamp.py
@@ -1,17 +1,17 @@
 """Data structures for timestamp information."""
 
-from . import common
+import logging
 
 import yorm
 
-log = common.logger(__name__)
+
+log = logging.getLogger(__name__)
 
 
 @yorm.attr(started=yorm.converters.Integer)
 @yorm.attr(stopped=yorm.converters.Integer)
 class Timestamp(yorm.converters.AttributeDictionary):
-
-    """A dictionary of last start and stop times."""
+    """Dictionary of last start and stop times."""
 
     def __init__(self, started=0, stopped=0):
         super().__init__()