+++ CHANGES.md
@@ -1,17 +1,23 @@
 # Version History
 
+## 0.3 (2015/11/14)
+
+- Added automatic daemon restart using `nohup`.
+- Moved `queued` to setting `properties.single_instance`.
+- Added `properties.auto_queue` to filter active applications.
+
 ## 0.2.1 (2015/09/05)
 
 - Fixed daemon warning to run using `nohup`
 
 ## 0.2 (2015/08/27)
 
-- Added '--daemon' option to run continuously.
-- Added 'edit' command to launch the settings file.
+- Added `--daemon` option to run continuously.
+- Added `edit` command to launch the settings file.
 
 ## 0.1.2 (2015/05/17)
 
-- Upgraded to YORM v0.4.
+- Updated to YORM v0.4.
 
 ## 0.1.1 (2015/03/19)